chatbot-master_sunhang/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-decouple==3.1
requests==2.22.0
soupsieve==1.9.2
urllib3==1.25.3
Werkzeug==0.15.4
